modified:   Java/src/dao/DAODPS.java 	new file:   Java/src/dao/DAONecessite.java 	new file:   Java/src/dao/DAOPossede.java 	modified:   Java/src/model/persistence/Possede.java 	new file:   Java/src/model/service/MngtBesoin.java 	new file:   Java/src/model/service/MngtCompetence.java 	new file:   Java/src/model/service/MngtJournee.java 	new file:   Java/src/model/service/MngtNecessite.java 	new file:   Java/src/model/service/MngtPossede.java 	new file:   Java/src/model/service/MngtSite.java 	new file:   Java/src/model/service/MngtSport.java

+++ Java/src/model/persistence/Possede.java
@@ -53,7 +53,7 @@ public class Possede {
      * Mutateur de l'attribut idSecouriste.
      * @param unIdSecouriste Le nouvel identifiant du secouriste possédant la compétence.
      */
-    public void setIdSecouriste(int unIdSecouriste) {
+    public void setIdSecouriste(long unIdSecouriste) {
         this.idSecouriste = unIdSecouriste;
     }
